+++ src/Symfony/Bundle/FrameworkBundle/Kernel/MicroKernelTrait.php
@@ -88,7 +88,7 @@ trait MicroKernelTrait
      */
     public function registerBundles(): iterable
     {
-        $contents = require $this->getProjectDir().'/config/bundles.php';
+        $contents = require $this->getBundlesPath();
         foreach ($contents as $class => $envs) {
             if ($envs[$this->environment] ?? $envs['all'] ?? false) {
                 yield new $class();
@@ -124,7 +124,7 @@ trait MicroKernelTrait
             $kernelDefinition->addTag('routing.route_loader');
 
             $container->addObjectResource($this);
-            $container->fileExists($this->getProjectDir().'/config/bundles.php');
+            $container->fileExists($this->getBundlesPath());
 
             try {
                 $configureContainer = new \ReflectionMethod($this, 'configureContainer');
@@ -202,5 +202,13 @@ trait MicroKernelTrait
         }
 
         return $collection;
+    }
+
+    /**
+     * Gets the path to the bundles configuration file.
+     */
+    private function getBundlesPath(): string
+    {
+        return $this->getProjectDir().'/config/bundles.php';
     }
 }